refactor(settings): move headers to include subdir
