+++ plugin/src/Caelestia/Settings/CMakeLists.txt
@@ -16,5 +16,5 @@ qml_module(caelestia-settings
 # Expose headers as settings/xxx.hpp
 set(settings_include_dir "${CMAKE_CURRENT_BINARY_DIR}/include")
 file(MAKE_DIRECTORY "${settings_include_dir}")
-file(CREATE_LINK "${CMAKE_CURRENT_SOURCE_DIR}" "${settings_include_dir}/settings" SYMBOLIC)
+file(CREATE_LINK "${CMAKE_CURRENT_SOURCE_DIR}/include" "${settings_include_dir}/settings" SYMBOLIC)
 target_include_directories(caelestia-settings PUBLIC "${settings_include_dir}")
