+++ plugin/src/Caelestia/Settings/changebatcher.cpp
@@ -1,4 +1,4 @@
-#include "changebatcher.hpp"
+#include "include/changebatcher.hpp"
 
 namespace caelestia::settings {
