+++ plugin/src/Caelestia/Settings/common.cpp
@@ -1,6 +1,6 @@
-#include "common.hpp"
+#include "include/common.hpp"
 
-#include "node.hpp"
+#include "include/node.hpp"
 
 namespace {
