+++ plugin/src/Caelestia/Settings/include/changebatcher.hpp
