+++ plugin/src/Caelestia/Settings/include/common.hpp
