+++ plugin/src/Caelestia/Settings/include/layerregistry.hpp
