+++ plugin/src/Caelestia/Settings/include/macros.hpp
