+++ plugin/src/Caelestia/Settings/include/objectnode.hpp
