+++ plugin/src/Caelestia/Settings/include/quarantine.hpp
