+++ plugin/src/Caelestia/Settings/include/schema.hpp
