+++ plugin/src/Caelestia/Settings/rootnode.cpp
@@ -1,4 +1,4 @@
-#include "rootnode.hpp"
+#include "include/rootnode.hpp"
 
 #include <qjsonobject.h>
